:recycle: Update interoperability initGenesisState
