+++ framework/src/modules/interoperability/sidechain/module.ts
@@ -25,8 +25,6 @@ import {
 	getTerminatedStateAccountRequestSchema,
 	getTerminatedOutboxAccountRequestSchema,
 } from '../schemas';
-import { GenesisBlockExecuteContext } from '../../../state_machine';
-import { initGenesisStateUtil } from '../utils';
 import {
 	chainAccountSchema,
 	allChainAccountsSchema,
@@ -141,10 +139,6 @@ export class SidechainInteroperabilityModule extends BaseInteroperabilityModule 
 				},
 			],
 		};
-	}
-
-	public async initGenesisState(context: GenesisBlockExecuteContext): Promise<void> {
-		await initGenesisStateUtil(context, this.stores);
 	}
 
 	// eslint-disable-next-line @typescript-eslint/require-await
