+++ framework/src/modules/interoperability/utils.ts
@@ -16,7 +16,6 @@ import { regularMerkleTree, sparseMerkleTree } from '@liskhq/lisk-tree';
 import { codec } from '@liskhq/lisk-codec';
 import { utils } from '@liskhq/lisk-cryptography';
 import { validator } from '@liskhq/lisk-validator';
-import { dataStructures } from '@liskhq/lisk-utils';
 import { NAME_REGEX } from '@liskhq/lisk-chain';
 import {
 	ActiveValidators,
@@ -27,48 +26,31 @@ import {
 	CrossChainUpdateTransactionParams,
 	ChainValidators,
 	InboxUpdate,
-	GenesisInteroperabilityInternalMethod,
 } from './types';
 import {
 	EMPTY_BYTES,
 	LIVENESS_LIMIT,
-	MAINCHAIN_ID_BUFFER,
 	MAX_CCM_SIZE,
-	MAX_NUM_VALIDATORS,
-	MAX_UINT64,
-	MODULE_NAME_INTEROPERABILITY,
 	SMT_KEY_LENGTH,
-	HASH_LENGTH,
-	TOKEN_ID_LSK,
 	CCMStatusCode,
 	CHAIN_ID_LENGTH,
 } from './constants';
 import {
 	ccmSchema,
-	genesisInteroperabilityInternalMethodSchema,
 	sidechainTerminatedCCMParamsSchema,
 	validatorsHashInputSchema,
 } from './schemas';
-import {
-	BlockHeader,
-	GenesisBlockExecuteContext,
-	VerificationResult,
-	VerifyStatus,
-} from '../../state_machine';
+import { BlockHeader, VerificationResult, VerifyStatus } from '../../state_machine';
 import { Certificate } from '../../engine/consensus/certificate_generation/types';
 import { certificateSchema } from '../../engine/consensus/certificate_generation/schema';
 import { CommandExecuteContext } from '../../state_machine/types';
 import { certificateToJSON } from './certificates';
 import { NamedRegistry } from '../named_registry';
 import { OutboxRootStore } from './stores/outbox_root';
-import { OwnChainAccountStore } from './stores/own_chain_account';
 import { ChannelDataStore } from './stores/channel_data';
 import { ChainValidatorsStore, calculateNewActiveValidators } from './stores/chain_validators';
 import { ChainAccountStore, ChainStatus } from './stores/chain_account';
-import { TerminatedOutboxAccount, TerminatedOutboxStore } from './stores/terminated_outbox';
-import { TerminatedStateStore } from './stores/terminated_state';
-import { RegisteredNamesStore } from './stores/registered_names';
-import { splitTokenID } from '../token/utils';
+import { TerminatedOutboxAccount } from './stores/terminated_outbox';
 
 interface CommonExecutionLogicArgs {
 	stores: NamedRegistry;
@@ -491,290 +473,6 @@ export const commonCCUExecutelogic = async (args: CommonExecutionLogicArgs) => {
 		);
 	}
 	await partnerChannelStore.set(context, chainIDBuffer, partnerChannelData);
-};
-
-export const initGenesisStateUtil = async (
-	ctx: GenesisBlockExecuteContext,
-	stores: NamedRegistry,
-) => {
-	const assetBytes = ctx.assets.getAsset(MODULE_NAME_INTEROPERABILITY);
-	if (!assetBytes) {
-		return;
-	}
-
-	const genesisStore = codec.decode<GenesisInteroperabilityInternalMethod>(
-		genesisInteroperabilityInternalMethodSchema,
-		assetBytes,
-	);
-	validator.validate(genesisInteroperabilityInternalMethodSchema, genesisStore);
-
-	const outboxRootStoreKeySet = new dataStructures.BufferSet();
-	const outboxRootStore = stores.get(OutboxRootStore);
-	for (const outboxRootData of genesisStore.outboxRootSubstore) {
-		if (outboxRootStoreKeySet.has(outboxRootData.storeKey)) {
-			throw new Error(
-				`Outbox root store key ${outboxRootData.storeKey.toString('hex')} is duplicated.`,
-			);
-		}
-		outboxRootStoreKeySet.add(outboxRootData.storeKey);
-		await outboxRootStore.set(ctx, outboxRootData.storeKey, outboxRootData.storeValue);
-	}
-
-	const ownChainAccountStore = stores.get(OwnChainAccountStore);
-	const ownChainAccount = await ownChainAccountStore.get(ctx, MAINCHAIN_ID_BUFFER);
-	const channelDataStoreKeySet = new dataStructures.BufferSet();
-	const channelDataStore = stores.get(ChannelDataStore);
-	for (const channelData of genesisStore.channelDataSubstore) {
-		if (channelDataStoreKeySet.has(channelData.storeKey)) {
-			throw new Error(
-				`Channel data store key ${channelData.storeKey.toString('hex')} is duplicated.`,
-			);
-		}
-		channelDataStoreKeySet.add(channelData.storeKey);
-
-		const channel = channelData.storeValue;
-		const chainID = splitTokenID(channel.messageFeeTokenID)[0];
-
-		if (
-			!channel.messageFeeTokenID.equals(TOKEN_ID_LSK) && // corresponding to the LSK token
-			!chainID.equals(channelData.storeKey) && // Token.getChainID(channel.messageFeeTokenID) must be equal to channelData.storeKey
-			!chainID.equals(ownChainAccount.chainID) // the message fee token must be a native token of either chains
-		) {
-			throw new Error(
-				`messageFeeTokenID corresponding to the channel data store key ${channelData.storeKey.toString(
-					'hex',
-				)} is not valid.`,
-			);
-		}
-
-		await channelDataStore.set(ctx, channelData.storeKey, channelData.storeValue);
-	}
-
-	const chainValidatorsStoreKeySet = new dataStructures.BufferSet();
-	const chainValidatorsStore = stores.get(ChainValidatorsStore);
-	for (const chainValidators of genesisStore.chainValidatorsSubstore) {
-		if (chainValidatorsStoreKeySet.has(chainValidators.storeKey)) {
-			throw new Error(
-				`Chain validators store key ${chainValidators.storeKey.toString('hex')} is duplicated.`,
-			);
-		}
-		chainValidatorsStoreKeySet.add(chainValidators.storeKey);
-
-		const { activeValidators, certificateThreshold } = chainValidators.storeValue;
-		if (activeValidators.length < 1 || activeValidators.length > MAX_NUM_VALIDATORS) {
-			throw new Error(
-				`Active validators must have at least 1 element and at most ${MAX_NUM_VALIDATORS} elements.`,
-			);
-		}
-
-		let totalWeight = BigInt(0);
-		for (let j = 0; j < activeValidators.length; j += 1) {
-			const activeValidator = activeValidators[j];
-
-			const { blsKey } = activeValidator;
-			if (j < activeValidators.length - 1 && blsKey.compare(activeValidators[j + 1].blsKey) >= 0) {
-				throw new Error(
-					'Active validators must be ordered lexicographically by blsKey property and pairwise distinct.',
-				);
-			}
-
-			const { bftWeight } = activeValidator;
-			totalWeight += bftWeight;
-		}
-
-		if (totalWeight > MAX_UINT64) {
-			throw new Error(
-				'The total BFT weight of all active validators has to be less than or equal to MAX_UINT64.',
-			);
-		}
-
-		const checkBftWeightValue = totalWeight / BigInt(3) + BigInt(1);
-		if (checkBftWeightValue > totalWeight || checkBftWeightValue > certificateThreshold) {
-			throw new Error('The total BFT weight of all active validators is not valid.');
-		}
-
-		await chainValidatorsStore.set(ctx, chainValidators.storeKey, chainValidators.storeValue);
-	}
-
-	const chainDataStoreKeySet = new dataStructures.BufferSet();
-	const chainDataStore = stores.get(ChainAccountStore);
-	let isAnotherSidechainAccount = 0;
-	for (const chainData of genesisStore.chainDataSubstore) {
-		const chainDataStoreKey = chainData.storeKey;
-		if (chainDataStoreKeySet.has(chainDataStoreKey)) {
-			throw new Error(`Chain data store key ${chainDataStoreKey.toString('hex')} is duplicated.`);
-		}
-		chainDataStoreKeySet.add(chainDataStoreKey);
-
-		const chainAccountStatus = chainData.storeValue.status;
-		if (chainAccountStatus === ChainStatus.TERMINATED) {
-			if (outboxRootStoreKeySet.has(chainDataStoreKey)) {
-				throw new Error('Outbox root store cannot have entry for a terminated chain account.');
-			}
-			if (
-				!channelDataStoreKeySet.has(chainDataStoreKey) ||
-				!chainValidatorsStoreKeySet.has(chainDataStoreKey)
-			) {
-				throw new Error(
-					`Chain data store key ${chainDataStoreKey.toString(
-						'hex',
-					)} missing in some or all of channel data and chain validators stores.`,
-				);
-			}
-		}
-		if (
-			!outboxRootStoreKeySet.has(chainDataStoreKey) ||
-			!channelDataStoreKeySet.has(chainDataStoreKey) ||
-			!chainValidatorsStoreKeySet.has(chainDataStoreKey)
-		) {
-			throw new Error(
-				`Chain data store key ${chainDataStoreKey.toString(
-					'hex',
-				)} missing in some or all of outbox root, channel data and chain validators stores.`,
-			);
-		}
-
-		if (
-			!(
-				chainDataStoreKey.equals(ownChainAccount.chainID) ||
-				chainDataStoreKey.equals(MAINCHAIN_ID_BUFFER)
-			)
-		) {
-			isAnotherSidechainAccount = 1;
-		}
-
-		await chainDataStore.set(ctx, chainData.storeKey, chainData.storeValue);
-	}
-
-	if (
-		isAnotherSidechainAccount &&
-		!(
-			chainDataStoreKeySet.has(MAINCHAIN_ID_BUFFER) &&
-			chainDataStoreKeySet.has(ownChainAccount.chainID)
-		)
-	) {
-		throw new Error(
-			'If a chain account for another sidechain is present, then a chain account for the mainchain must be present, as well as the own chain account.',
-		);
-	}
-
-	for (const storeKey of outboxRootStoreKeySet) {
-		if (!chainDataStoreKeySet.has(storeKey)) {
-			throw new Error(
-				`Outbox root store key ${storeKey.toString('hex')} is missing in chain data store.`,
-			);
-		}
-	}
-
-	for (const storeKey of channelDataStoreKeySet) {
-		if (!chainDataStoreKeySet.has(storeKey)) {
-			throw new Error(
-				`Channel data store key ${storeKey.toString('hex')} is missing in chain data store.`,
-			);
-		}
-	}
-
-	for (const storeKey of chainValidatorsStoreKeySet) {
-		if (!chainDataStoreKeySet.has(storeKey)) {
-			throw new Error(
-				`Chain validators store key ${storeKey.toString('hex')} is missing in chain data store.`,
-			);
-		}
-	}
-
-	const ownChainDataStoreKeySet = new dataStructures.BufferSet();
-	for (const ownChainData of genesisStore.ownChainDataSubstore) {
-		if (ownChainDataStoreKeySet.has(ownChainData.storeKey)) {
-			throw new Error(
-				`Own chain data store key ${ownChainData.storeKey.toString('hex')} is duplicated.`,
-			);
-		}
-		ownChainDataStoreKeySet.add(ownChainData.storeKey);
-
-		await ownChainAccountStore.set(ctx, ownChainData.storeKey, ownChainData.storeValue);
-	}
-
-	const terminatedOutboxStoreKeySet = new dataStructures.BufferSet();
-	const terminatedOutboxStore = stores.get(TerminatedOutboxStore);
-	for (const terminatedOutbox of genesisStore.terminatedOutboxSubstore) {
-		if (terminatedOutboxStoreKeySet.has(terminatedOutbox.storeKey)) {
-			throw new Error(
-				`Terminated outbox store key ${terminatedOutbox.storeKey.toString('hex')} is duplicated.`,
-			);
-		}
-		terminatedOutboxStoreKeySet.add(terminatedOutbox.storeKey);
-
-		await terminatedOutboxStore.set(ctx, terminatedOutbox.storeKey, terminatedOutbox.storeValue);
-	}
-
-	const terminatedStateStoreKeySet = new dataStructures.BufferSet();
-	const terminatedStateStore = stores.get(TerminatedStateStore);
-	for (const terminatedState of genesisStore.terminatedStateSubstore) {
-		const terminatedStateStoreKey = terminatedState.storeKey;
-		if (terminatedStateStoreKeySet.has(terminatedStateStoreKey)) {
-			throw new Error(
-				`Terminated state store key ${terminatedStateStoreKey.toString('hex')} is duplicated.`,
-			);
-		}
-		terminatedStateStoreKeySet.add(terminatedStateStoreKey);
-
-		const terminatedStateStoreValue = terminatedState.storeValue;
-		if (terminatedStateStoreValue.initialized === false) {
-			if (terminatedOutboxStoreKeySet.has(terminatedStateStoreKey)) {
-				throw new Error(
-					`Uninitialized account associated with terminated state store key ${terminatedStateStoreKey.toString(
-						'hex',
-					)} cannot be present in terminated outbox store.`,
-				);
-			}
-			if (
-				!terminatedStateStoreValue.stateRoot.equals(EMPTY_BYTES) ||
-				terminatedStateStoreValue.mainchainStateRoot?.length !== 32
-			) {
-				throw new Error(
-					`For the uninitialized account associated with terminated state store key ${terminatedStateStoreKey.toString(
-						'hex',
-					)} the stateRoot must be set to empty bytes and mainchainStateRoot to a 32-bytes value.`,
-				);
-			}
-		} else if (terminatedStateStoreValue.initialized === true) {
-			if (
-				terminatedStateStoreValue.stateRoot.length !== HASH_LENGTH ||
-				terminatedStateStoreValue.mainchainStateRoot?.length !== HASH_LENGTH
-			) {
-				throw new Error(
-					`For the initialized account associated with terminated state store key ${terminatedStateStoreKey.toString(
-						'hex',
-					)} the mainchainStateRoot must be set to a 32-bytes value and stateRoot to a 32-bytes value.`,
-				);
-			}
-		}
-
-		await terminatedStateStore.set(ctx, terminatedState.storeKey, terminatedState.storeValue);
-	}
-
-	for (const storeKey of terminatedOutboxStoreKeySet) {
-		if (!terminatedStateStoreKeySet.has(storeKey)) {
-			throw new Error(
-				`Terminated outbox store key ${storeKey.toString(
-					'hex',
-				)} missing in terminated state store.`,
-			);
-		}
-	}
-
-	const registeredNamesStoreKeySet = new dataStructures.BufferSet();
-	const registeredNamesStore = stores.get(RegisteredNamesStore);
-	for (const registeredNames of genesisStore.registeredNamesSubstore) {
-		if (registeredNamesStoreKeySet.has(registeredNames.storeKey)) {
-			throw new Error(
-				`Registered names store key ${registeredNames.storeKey.toString('hex')} is duplicated.`,
-			);
-		}
-		registeredNamesStoreKeySet.add(registeredNames.storeKey);
-
-		await registeredNamesStore.set(ctx, registeredNames.storeKey, registeredNames.storeValue);
-	}
 };
 
 export const chainAccountToJSON = (chainAccount: ChainAccount) => {
