+++ framework/test/unit/modules/interoperability/utils.spec.ts
@@ -16,7 +16,6 @@ import { utils } from '@liskhq/lisk-cryptography';
 import { codec } from '@liskhq/lisk-codec';
 import * as cryptography from '@liskhq/lisk-cryptography';
 import * as merkleTree from '@liskhq/lisk-tree';
-import { BlockAssets } from '@liskhq/lisk-chain';
 import { MainchainInteroperabilityModule, VerifyStatus } from '../../../../src';
 import {
 	CCMStatusCode,
@@ -25,17 +24,10 @@ import {
 	EMPTY_BYTES,
 	HASH_LENGTH,
 	LIVENESS_LIMIT,
-	MAINCHAIN_ID_BUFFER,
 	MAX_CCM_SIZE,
-	MAX_NUM_VALIDATORS,
-	MAX_UINT64,
 	MODULE_NAME_INTEROPERABILITY,
-	TOKEN_ID_LSK,
 } from '../../../../src/modules/interoperability/constants';
-import {
-	ccmSchema,
-	genesisInteroperabilityInternalMethodSchema,
-} from '../../../../src/modules/interoperability/schemas';
+import { ccmSchema } from '../../../../src/modules/interoperability/schemas';
 import {
 	ChainAccount,
 	ChannelData,
@@ -53,7 +45,6 @@ import {
 	commonCCUExecutelogic,
 	computeValidatorsHash,
 	getIDAsKeyForStore,
-	initGenesisStateUtil,
 	validateFormat,
 	verifyLivenessConditionForRegisteredChains,
 } from '../../../../src/modules/interoperability/utils';
@@ -61,20 +52,12 @@ import { certificateSchema } from '../../../../src/engine/consensus/certificate_
 import { Certificate } from '../../../../src/engine/consensus/certificate_generation/types';
 import { PrefixedStateReadWriter } from '../../../../src/state_machine/prefixed_state_read_writer';
 import { InMemoryPrefixedStateDB } from '../../../../src/testing/in_memory_prefixed_state';
-import { createGenesisBlockContext } from '../../../../src/testing';
 import {
 	ChainAccountStore,
 	ChainStatus,
 } from '../../../../src/modules/interoperability/stores/chain_account';
-import { RegisteredNamesStore } from '../../../../src/modules/interoperability/stores/registered_names';
 import { ChainValidatorsStore } from '../../../../src/modules/interoperability/stores/chain_validators';
-import { TerminatedStateStore } from '../../../../src/modules/interoperability/stores/terminated_state';
-import { TerminatedOutboxStore } from '../../../../src/modules/interoperability/stores/terminated_outbox';
-import { OutboxRootStore } from '../../../../src/modules/interoperability/stores/outbox_root';
 import { ChannelDataStore } from '../../../../src/modules/interoperability/stores/channel_data';
-import { OwnChainAccountStore } from '../../../../src/modules/interoperability/stores/own_chain_account';
-import { createStoreGetter } from '../../../../src/testing/utils';
-import { CHAIN_ID_LENGTH } from '../../../../src/modules/token/constants';
 
 jest.mock('@liskhq/lisk-cryptography', () => ({
 	...jest.requireActual('@liskhq/lisk-cryptography'),
@@ -974,901 +957,6 @@ describe('Utils', () => {
 				partnerChannelData.inbox.root,
 			);
 			expect(calculateRootFromRightWitness).not.toHaveBeenCalled();
-		});
-	});
-
-	describe('initGenesisStateUtil', () => {
-		const { getRandomBytes } = cryptography.utils;
-		const timestamp = 2592000 * 100;
-		const chainAccount = {
-			name: 'account1',
-			chainID: Buffer.alloc(CHAIN_ID_LENGTH),
-			lastCertificate: {
-				height: 567467,
-				timestamp: timestamp - 500000,
-				stateRoot: Buffer.alloc(HASH_LENGTH),
-				validatorsHash: Buffer.alloc(HASH_LENGTH),
-			},
-			status: 2739,
-		};
-		const sidechainChainAccount = {
-			name: 'sidechain1',
-			chainID: Buffer.alloc(CHAIN_ID_LENGTH),
-			lastCertificate: {
-				height: 10,
-				stateRoot: utils.getRandomBytes(32),
-				timestamp: 100,
-				validatorsHash: utils.getRandomBytes(32),
-			},
-			status: ChainStatus.TERMINATED,
-		};
-		const ownChainAccount = {
-			name: 'mainchain',
-			chainID: MAINCHAIN_ID_BUFFER,
-			nonce: BigInt('0'),
-		};
-		const channelData = {
-			inbox: {
-				appendPath: [Buffer.alloc(HASH_LENGTH), Buffer.alloc(HASH_LENGTH)],
-				root: cryptography.utils.getRandomBytes(HASH_LENGTH),
-				size: 18,
-			},
-			messageFeeTokenID: TOKEN_ID_LSK,
-			outbox: {
-				appendPath: [Buffer.alloc(HASH_LENGTH), Buffer.alloc(HASH_LENGTH)],
-				root: cryptography.utils.getRandomBytes(HASH_LENGTH),
-				size: 18,
-			},
-			partnerChainOutboxRoot: cryptography.utils.getRandomBytes(HASH_LENGTH),
-		};
-		const outboxRoot = { root: getRandomBytes(HASH_LENGTH) };
-		const validatorsHashInput = {
-			activeValidators: [
-				{
-					blsKey: Buffer.from(
-						'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-						'hex',
-					),
-					bftWeight: BigInt(10),
-				},
-				{
-					blsKey: Buffer.from(
-						'4c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-						'hex',
-					),
-					bftWeight: BigInt(10),
-				},
-			],
-			certificateThreshold: BigInt(10),
-		};
-		const terminatedStateAccount = {
-			stateRoot: sidechainChainAccount.lastCertificate.stateRoot,
-			mainchainStateRoot: Buffer.alloc(HASH_LENGTH),
-			initialized: true,
-		};
-		const terminatedOutboxAccount = {
-			outboxRoot: getRandomBytes(HASH_LENGTH),
-			outboxSize: 1,
-			partnerChainInboxSize: 1,
-		};
-		const registeredNameId = { chainID: Buffer.alloc(CHAIN_ID_LENGTH) };
-		const registeredChainId = { chainID: Buffer.alloc(CHAIN_ID_LENGTH) };
-		const validData = {
-			outboxRootSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: outboxRoot },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: outboxRoot },
-			],
-			chainDataSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: chainAccount },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: chainAccount },
-			],
-			channelDataSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: channelData },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: channelData },
-			],
-			chainValidatorsSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: validatorsHashInput },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-			],
-			ownChainDataSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: ownChainAccount },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: ownChainAccount },
-			],
-			terminatedStateSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: terminatedStateAccount },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedStateAccount },
-			],
-			terminatedOutboxSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: terminatedOutboxAccount },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedOutboxAccount },
-			],
-			registeredNamesSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: registeredNameId },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: registeredNameId },
-			],
-			registeredChainIDsSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: registeredChainId },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: registeredChainId },
-			],
-		};
-
-		const invalidData = {
-			...validData,
-			outboxRootSubstore: [
-				{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: { root: getRandomBytes(37) } },
-				{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: { root: getRandomBytes(5) } },
-			],
-		};
-
-		let channelDataSubstore: ChannelDataStore;
-		let outboxRootSubstore: OutboxRootStore;
-		let terminatedOutboxSubstore: TerminatedOutboxStore;
-		let stateStore: PrefixedStateReadWriter;
-		let chainDataSubstore: ChainAccountStore;
-		let terminatedStateSubstore: TerminatedStateStore;
-		let chainValidatorsSubstore: ChainValidatorsStore;
-		let ownChainDataSubstore: OwnChainAccountStore;
-		let registeredNamesSubstore: RegisteredNamesStore;
-
-		beforeEach(async () => {
-			stateStore = new PrefixedStateReadWriter(new InMemoryPrefixedStateDB());
-			ownChainDataSubstore = interopMod.stores.get(OwnChainAccountStore);
-			await ownChainDataSubstore.set(
-				createStoreGetter(stateStore),
-				MAINCHAIN_ID_BUFFER,
-				ownChainAccount,
-			);
-			channelDataSubstore = interopMod.stores.get(ChannelDataStore);
-			chainValidatorsSubstore = interopMod.stores.get(ChainValidatorsStore);
-			outboxRootSubstore = interopMod.stores.get(OutboxRootStore);
-			terminatedOutboxSubstore = interopMod.stores.get(TerminatedOutboxStore);
-			chainDataSubstore = interopMod.stores.get(ChainAccountStore);
-			terminatedStateSubstore = interopMod.stores.get(TerminatedStateStore);
-			registeredNamesSubstore = interopMod.stores.get(RegisteredNamesStore);
-		});
-
-		it('should not throw error if asset does not exist', async () => {
-			const context = createGenesisBlockContext({ stateStore }).createInitGenesisStateContext();
-			jest.spyOn(context, 'getStore');
-
-			await expect(initGenesisStateUtil(context, interopMod.stores)).toResolve();
-			expect(context.getStore).not.toHaveBeenCalled();
-		});
-
-		it('should throw if the asset object is invalid', async () => {
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, invalidData);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if outbox root store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				outboxRootSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: outboxRoot },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: outboxRoot },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if chain data store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				chainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: chainAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: chainAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if channel data store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				channelDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: channelData },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: channelData },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if chain validators store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if own chain store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				ownChainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: ownChainAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: ownChainAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if terminated state store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				terminatedStateSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedStateAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedStateAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if terminated outbox store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				terminatedOutboxSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedOutboxAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedOutboxAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if registered names store key is duplicated', async () => {
-			const validData1 = {
-				...validData,
-				registeredNamesSubstore: [
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: registeredNameId },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: registeredNameId },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in chain data substore is missing in outbox root substore and the corresponding chain account is not inactive', async () => {
-			const validData1 = {
-				...validData,
-				outboxRootSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: outboxRoot },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: outboxRoot },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in chain data substore is present in outbox root substore but the corresponding chain account is inactive', async () => {
-			const validData1 = {
-				...validData,
-				chainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: { ...chainAccount, status: 2 } },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: chainAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in chain data substore is missing in channel data substore', async () => {
-			const validData1 = {
-				...validData,
-				channelDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: channelData },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: channelData },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in chain data substore is missing in chain validators substore', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: validatorsHashInput },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in outbox data substore is missing in chain data substore', async () => {
-			const validData1 = {
-				...validData,
-				chainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: chainAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 1]), storeValue: chainAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in channel data substore is missing in chain data substore', async () => {
-			const validData1 = {
-				...validData,
-				chainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: chainAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 1]), storeValue: chainAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in chain validators substore is missing in chain data substore', async () => {
-			const validData1 = {
-				...validData,
-				chainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: chainAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 1]), storeValue: chainAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in terminated outbox substore is missing in the terminated state substore', async () => {
-			const validData1 = {
-				...validData,
-				terminatedStateSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: terminatedStateAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedStateAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in terminated state substore is present in the terminated outbox substore but the property initialized is set to false', async () => {
-			const validData1 = {
-				...validData,
-				terminatedStateSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: { ...terminatedStateAccount, initialized: false },
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedStateAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in terminated state substore has the property initialized set to false but stateRoot is not set to empty bytes and mainchainStateRoot not set to a 32-bytes value', async () => {
-			const validData1 = {
-				...validData,
-				terminatedStateSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 0]),
-						storeValue: { ...terminatedStateAccount, initialized: false },
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedStateAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some store key in terminated state substore has the property initialized set to true but mainchainStateRoot is not set to empty bytes and stateRoot not set to a 32-bytes value', async () => {
-			const validData1 = {
-				...validData,
-				terminatedStateSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: {
-							...terminatedStateAccount,
-							initialized: true,
-							mainchainStateRoot: getRandomBytes(32),
-							stateRoot: EMPTY_BYTES,
-						},
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: terminatedStateAccount },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if active validators have less than 1 element', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: { ...validatorsHashInput, activeValidators: [] },
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if active validators have more than MAX_NUM_VALIDATORS elements', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: {
-							...validatorsHashInput,
-							activeValidators: new Array(MAX_NUM_VALIDATORS + 1).fill(0),
-						},
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if active validators are not ordered lexicographically by blsKey', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: {
-							...validatorsHashInput,
-							activeValidators: [
-								{
-									blsKey: Buffer.from(
-										'4c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(10),
-								},
-								{
-									blsKey: Buffer.from(
-										'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(10),
-								},
-							],
-						},
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some active validators have blsKey which is not 48 bytes', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: {
-							...validatorsHashInput,
-							activeValidators: [
-								{
-									blsKey: getRandomBytes(21),
-									bftWeight: BigInt(10),
-								},
-								{
-									blsKey: Buffer.from(
-										'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(10),
-								},
-							],
-						},
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some active validators have blsKey which is not pairwise distinct', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: {
-							...validatorsHashInput,
-							activeValidators: [
-								{
-									blsKey: Buffer.from(
-										'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(10),
-								},
-								{
-									blsKey: Buffer.from(
-										'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(10),
-								},
-							],
-						},
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if some active validators have bftWeight which is not a positive integer', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: {
-							...validatorsHashInput,
-							activeValidators: [
-								{
-									blsKey: Buffer.from(
-										'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(-1),
-								},
-								{
-									blsKey: Buffer.from(
-										'4c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(10),
-								},
-							],
-						},
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if total bft weight of active validators is greater than MAX_UINT64', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: {
-							...validatorsHashInput,
-							activeValidators: [
-								{
-									blsKey: Buffer.from(
-										'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(MAX_UINT64),
-								},
-								{
-									blsKey: Buffer.from(
-										'4c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-										'hex',
-									),
-									bftWeight: BigInt(10),
-								},
-							],
-						},
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if total bft weight of active validators is less than the value check', async () => {
-			const validatorsHashInput1 = {
-				...validatorsHashInput,
-				activeValidators: [
-					{
-						blsKey: Buffer.from(
-							'3c1e6f29e3434f816cd6697e56cc54bc8d80927bf65a1361b383aa338cd3f63cbf82ce801b752cb32f8ecb3f8cc16835',
-							'hex',
-						),
-						bftWeight: BigInt(0),
-					},
-				],
-			};
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: validatorsHashInput1 },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput1 },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if certificateThreshold is less than the value check', async () => {
-			const validData1 = {
-				...validData,
-				chainValidatorsSubstore: [
-					{
-						storeKey: Buffer.from([0, 0, 0, 1]),
-						storeValue: { ...validatorsHashInput, certificateThreshold: BigInt(1) },
-					},
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if a chain account for another sidechain is present but chain account for mainchain is not present', async () => {
-			const validData1 = {
-				...validData,
-				chainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: chainAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: chainAccount },
-				],
-				outboxRootSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: outboxRoot },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: outboxRoot },
-				],
-				channelDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: channelData },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: channelData },
-				],
-				chainValidatorsSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: validatorsHashInput },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should throw if a chain account for another sidechain is present but chain account for ownchain is not present', async () => {
-			const validData1 = {
-				...validData,
-				chainDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: chainAccount },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: chainAccount },
-				],
-				outboxRootSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: outboxRoot },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: outboxRoot },
-				],
-				channelDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: channelData },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: channelData },
-				],
-				chainValidatorsSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 0]), storeValue: validatorsHashInput },
-					{ storeKey: Buffer.from([0, 0, 1, 0]), storeValue: validatorsHashInput },
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).rejects.toThrow();
-		});
-
-		it('should not throw if some chain id corresponding to message fee token id of a channel is not 1 but is corresponding native token id of either chains', async () => {
-			const validData1 = {
-				...validData,
-				channelDataSubstore: [
-					{ storeKey: Buffer.from([0, 0, 0, 1]), storeValue: channelData },
-					{
-						storeKey: Buffer.from([0, 0, 1, 0]),
-						storeValue: {
-							...channelData,
-							messageFeeTokenID: TOKEN_ID_LSK,
-						},
-					},
-				],
-			};
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData1);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-			await expect(initGenesisStateUtil(context, interopMod.stores)).toResolve();
-		});
-
-		it('should create all the corresponding entries in the interoperability module state for every substore for valid input', async () => {
-			const encodedAsset = codec.encode(genesisInteroperabilityInternalMethodSchema, validData);
-			const context = createGenesisBlockContext({
-				stateStore,
-				assets: new BlockAssets([{ module: MODULE_NAME_INTEROPERABILITY, data: encodedAsset }]),
-			}).createInitGenesisStateContext();
-
-			await expect(initGenesisStateUtil(context, interopMod.stores)).toResolve();
-
-			channelDataSubstore = interopMod.stores.get(ChannelDataStore);
-			chainValidatorsSubstore = interopMod.stores.get(ChainValidatorsStore);
-			outboxRootSubstore = interopMod.stores.get(OutboxRootStore);
-			terminatedOutboxSubstore = interopMod.stores.get(TerminatedOutboxStore);
-			chainDataSubstore = interopMod.stores.get(ChainAccountStore);
-			terminatedStateSubstore = interopMod.stores.get(TerminatedStateStore);
-			registeredNamesSubstore = interopMod.stores.get(RegisteredNamesStore);
-			ownChainDataSubstore = interopMod.stores.get(OwnChainAccountStore);
-
-			for (const data of validData.chainDataSubstore) {
-				await expect(
-					chainDataSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
-			for (const data of validData.chainValidatorsSubstore) {
-				await expect(
-					chainValidatorsSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
-			for (const data of validData.outboxRootSubstore) {
-				await expect(
-					outboxRootSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
-			for (const data of validData.terminatedOutboxSubstore) {
-				await expect(
-					terminatedOutboxSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
-			for (const data of validData.channelDataSubstore) {
-				await expect(
-					channelDataSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
-			for (const data of validData.terminatedStateSubstore) {
-				await expect(
-					terminatedStateSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
-			for (const data of validData.registeredNamesSubstore) {
-				await expect(
-					registeredNamesSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
-			for (const data of validData.ownChainDataSubstore) {
-				await expect(
-					ownChainDataSubstore.has(createStoreGetter(stateStore), data.storeKey),
-				).resolves.toBeTrue();
-			}
 		});
 	});
